# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Celljevity Longevity OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [NEW, CONTACTED, QUALIFIED, CONVERTED, LOST]
        - name: source
          in: query
          schema:
            type: string
            enum: [WEBSITE, REFERRAL, PARTNER, EVENT, SOCIAL_MEDIA, DIRECT, OTHER]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{leadId}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead details
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadInput"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /leads/{leadId}/convert:
    post:
      operationId: convertLead
      tags: [leads]
      summary: Convert lead to patient
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead converted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  patient:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      celljevityId:
                        type: string
                      userId:
                        type: string
                        format: uuid
                  temporaryPassword:
                    type: string
...
  /patients/{patientId}/intake:
    get:
      operationId: getIntakeForm
      tags: [intake]
      summary: Get intake form
      parameters:
        - $ref: "#/components/parameters/PatientId"
      responses:
        "200":
          description: Intake form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeForm"
    post:
      operationId: createIntakeForm
      tags: [intake]
      summary: Create intake form
      parameters:
        - $ref: "#/components/parameters/PatientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntakeFormData"
      responses:
        "201":
          description: Intake form created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeForm"
    patch:
      operationId: updateIntakeForm
      tags: [intake]
      summary: Update intake form
      parameters:
        - $ref: "#/components/parameters/PatientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntakeFormData"
      responses:
        "200":
          description: Intake form updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeForm"
  /patients/{patientId}/intake/complete:
    post:
      operationId: completeIntakeForm
      tags: [intake]
      summary: Complete intake and advance journey
      parameters:
        - $ref: "#/components/parameters/PatientId"
      responses:
        "200":
          description: Intake completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeForm"
...
components:
  parameters:
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        source:
          type: string
          enum: [WEBSITE, REFERRAL, PARTNER, EVENT, SOCIAL_MEDIA, DIRECT, OTHER]
        status:
          type: string
          enum: [NEW, CONTACTED, QUALIFIED, CONVERTED, LOST]
        notes:
          type: string
          nullable: true
        assignedTo:
          type: string
          format: uuid
          nullable: true
        convertedPatientId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateLeadInput:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: string
          enum: [WEBSITE, REFERRAL, PARTNER, EVENT, SOCIAL_MEDIA, DIRECT, OTHER]
        notes:
          type: string
    UpdateLeadInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        source:
          type: string
          enum: [WEBSITE, REFERRAL, PARTNER, EVENT, SOCIAL_MEDIA, DIRECT, OTHER]
        status:
          type: string
          enum: [NEW, CONTACTED, QUALIFIED, CONVERTED, LOST]
        notes:
          type: string
        assignedTo:
          type: string
          format: uuid
...
    IntakeForm:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        personalProfile:
          type: object
          nullable: true
        medicalHistory:
          type: object
          nullable: true
        consentData:
          type: object
          nullable: true
        isComplete:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, patientId, isComplete, createdAt, updatedAt]
    IntakeFormData:
      type: object
      properties:
        personalProfile:
          type: object
        medicalHistory:
          type: object
        consentData:
          type: object
...